requirements.txt
django == 5.0.6
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.3.1
drf-yasg == 1.21.7
psycopg2-binary == 2.9.9
python-dotenv == 1.0.1
redis == 5.0.7
celery == 5.4.0
django-celery-beat == 2.6.0
pillow == 10.3.0
ipython == 8.25.0